Extract MongoMapper formmatter
