+++ lib/awesome_print/formatters/mongo_mapper_bson_id.rb
@@ -1,0 +1,10 @@
+module AwesomePrint
+  module Formatters
+    class MongoMapperBsonId < Base
+
+      def call
+        object.inspect
+      end
+    end
+  end
+end
